refactor(data): model 패키지와 data 패키지 분리 및 역할에 맞는 패키지로 구조 이동
- model패키지는 서비스와 같은 역할을 하는 패키지로 설정
- data패키지는 상태를 갖는 데이터 역할 패키지로 설정
