+++ src/lotto/data/Lotto.js
