+++ src/lotto/data/WinningResult.js
