+++ src/lotto/model/RateCalculator.js
